[Fix] Login & Test Bug fix

+++ app/src/main/java/com/example/konwnow/data/remote/dto/Users.kt
@@ -19,7 +19,7 @@ class Users{
 
     data class DataList(
         @SerializedName("correctPercentage")
-        val correctPercentage: Int,
+        val correctPercentage: Double,
         @SerializedName("examCount")
         val examCount: Int,
         @SerializedName("id")
